+++ lib/inspec/objects.rb
@@ -1,5 +1,5 @@
 module Inspec
-  autoload :Input, "inspec/objects/input"
+  # TODO: these should be namespaced in Objects
   autoload :Tag, "inspec/objects/tag"
   autoload :Control, "inspec/objects/control"
   autoload :Describe, "inspec/objects/describe"
@@ -10,3 +10,5 @@ module Inspec
   autoload :Test, "inspec/objects/test"
   autoload :Value, "inspec/objects/value"
 end
+
+require "inspec/objects/input" # already defined so you can't autoload
