+++ test/unit/inputs/validation_test.rb
@@ -1,5 +1,5 @@
 require "helper"
-require "inspec/objects/input"
+require "inspec/input"
 
 describe "type validation" do
   let(:opts) { {} }
